Refactor Parser and Scanner Definitions; Optimize Imports; Re-organize Cmake File; Make g++ Cry Less
There were a lot of files that needed to be cleaned up; however, the biggest problem was that I had forgotten to include Driver.h in my flex file. Moreover, a certain important static variable needed to be instantiated outside of its class definition in order to be usable where it ought to be. I'm no longer held back by an undefined symbol error; now, it's innumerably more, albeit easier-to-fix, smaller errors. Progress!