Implement correct token construction def's in scanner.l; refactor register pool into vector member variable of Register class.
Token construction is no longer throwing compile-time errors, and linking is now successful upon running the `make` command. The next commit will inevitably be me tuning the Driver of the program to ensure running the program with valid parameters will act as expected. Once this testing is finished, it'll be on to conditional statements for this project!