First Little Bit of Function-ality
Laying the groundwork for Functions and the functions they contain. I decided that a `struct` was a more appropriate structure for a Function and its member objects, as nothing incredibly special lives inside, and thus, may be public to the rest of the compiler (plus, I'm lazy and don't want to create getters and setters).